Rename(#2-Error-control): 패키지 구조 변경, close #2
