+++ src/main/java/com/_8attery/seesaw/exception/custom/InvalidRequestException.java
@@ -1,4 +1,4 @@
-package com._8attery.seesaw.exception;
+package com._8attery.seesaw.exception.custom;
 
 public class InvalidRequestException extends RuntimeException {
     public InvalidRequestException(String message) {
