+++ src/main/java/com/_8attery/seesaw/exception/custom/ResourceNotFoundException.java
@@ -1,4 +1,4 @@
-package com._8attery.seesaw.exception;
+package com._8attery.seesaw.exception.custom;
 
 public class ResourceNotFoundException extends RuntimeException {
     public ResourceNotFoundException(String message) {
