+++ src/main/java/com/_8attery/seesaw/exception/custom/UserUnauthorizedException.java
@@ -1,4 +1,4 @@
-package com._8attery.seesaw.exception;
+package com._8attery.seesaw.exception.custom;
 
 public class UserUnauthorizedException extends RuntimeException {
     public UserUnauthorizedException(String message) {
